Avoid code duplication in bid/place delegate tests.
Creates a base class that's inherited by both tests holding test cases that have the same expectations for the two.
